+++ back/src/main/java/com/ll/goohaeyou/domain/jobPost/jobPost/controller/InterestController.java
@@ -1,5 +1,6 @@
 package com.ll.goohaeyou.domain.jobPost.jobPost.controller;
 
+import com.ll.goohaeyou.domain.jobPost.jobPost.service.InterestService;
 import com.ll.goohaeyou.domain.jobPost.jobPost.service.JobPostService;
 import com.ll.goohaeyou.global.apiResponse.ApiResponse;
 import com.ll.goohaeyou.global.security.MemberDetails;
@@ -16,12 +17,13 @@ import org.springframework.web.bind.annotation.*;
 @Tag(name = "Interest", description = "관심 공고 API")
 public class InterestController {
     private final JobPostService jobPostService;
+    private final InterestService interestService;
 
     @PostMapping("/{id}/interest")
     @Operation(summary = "구인공고 관심 등록")
     public ApiResponse<Empty> interest(@AuthenticationPrincipal MemberDetails memberDetails,
                                        @PathVariable(name = "id") Long id) {
-        jobPostService.interest(memberDetails.getUsername(), id);
+        interestService.interest(memberDetails.getUsername(), id);
 
         return ApiResponse.noContent();
     }
@@ -30,7 +32,7 @@ public class InterestController {
     @Operation(summary = "구인공고 관심 제거")
     public ApiResponse<Empty> disinterest(@AuthenticationPrincipal MemberDetails memberDetails,
                                           @PathVariable(name = "id") Long id) {
-        jobPostService.disinterest(memberDetails.getUsername(), id);
+        interestService.disinterest(memberDetails.getUsername(), id);
 
         return ApiResponse.noContent();
     }
@@ -39,7 +41,7 @@ public class InterestController {
     @Operation(summary = "로그인한 유저의 해당 구인공고 관심 등록 여부")
     public ApiResponse<Boolean> isInterested(@AuthenticationPrincipal MemberDetails memberDetails,
                                              @PathVariable(name = "id") Long id) {
-        boolean result = jobPostService.isInterested(memberDetails.getUsername(), id);
+        boolean result = interestService.isInterested(memberDetails.getUsername(), id);
 
         return ApiResponse.ok(result);
     }
